Add an API to only run 1 BenchmarkCollection
API has the same format as we're used from unit tests
